feat：<行为树状态帧移植> 完成第一版，附带测试用例

+++ Unity/Assets/Hotfix/NKGMOBA/Battle/LockStepStateFrameSync/LSF_ComponentEvents.cs
@@ -20,7 +20,7 @@ namespace ET
             lsfComponent.HalfRTT = a.C2MPing % 2 == 0 ? a.C2MPing / 2 : a.C2MPing / 2 + 1;
             
             int targetAheadOfFrame =
-                TimeAndFrameConverter.Frame_Long2Frame(lsfComponent.HalfRTT) + (int)lsfComponent.BufferFrame;
+                (int)TimeAndFrameConverter.Frame_Long2Frame(lsfComponent.HalfRTT) + (int)lsfComponent.BufferFrame;
 
             lsfComponent.TargetAheadOfFrame =
                 targetAheadOfFrame > LSF_Component.AheadOfFrameMax
