+++ Unity/Assets/Model/NKGMOBA/Battle/LockStepStateFrameSync/Helper/TimeAndFrameConverter.cs
@@ -17,22 +17,22 @@ namespace ET
             return (time / 1000f);
         }
 
-        public static int Frame_Long2Frame(long time)
+        public static uint Frame_Long2Frame(long time)
         {
-            return Mathf.CeilToInt(((time * 1.0f) / GlobalDefine.FixedUpdateTargetDTTime_Long));
+            return (uint)Mathf.CeilToInt(((time * 1.0f) / GlobalDefine.FixedUpdateTargetDTTime_Long));
         }
 
-        public static int Frame_Float2Frame(float time)
+        public static uint Frame_Float2Frame(float time)
         {
-            return Mathf.CeilToInt(time / GlobalDefine.FixedUpdateTargetDTTime_Float);
+            return (uint)Mathf.CeilToInt(time / GlobalDefine.FixedUpdateTargetDTTime_Float);
         }
 
-        public static int Frame_Float2FrameWithHalfRTT(float time, long halfRTT)
+        public static uint Frame_Float2FrameWithHalfRTT(float time, long halfRTT)
         {
             return Frame_Long2Frame(((long) (time * 1000) + halfRTT));
         }
         
-        public static int Frame_Long2FrameWithHalfRTT(long time, long halfRTT)
+        public static uint Frame_Long2FrameWithHalfRTT(long time, long halfRTT)
         {
             return Frame_Long2Frame(time + halfRTT);
         }
